requirements.txt
Flask
Flask-RESTful
Flask-PyMongo
Flask-JWT-Extended
